Audit + CleanUp
git-svn-id: https://cchlib.googlecode.com/svn/trunk@8245 bc589049-105d-1923-a3ce-6fcb24364ed1

+++ cchlib-core-deprecated/src/test/java/cx/ath/choisnet/util/iterator/ArrayIteratorTest.java
@@ -53,7 +53,7 @@ public class ArrayIteratorTest
         Integer                firstValue = null;
 
         while( iter.hasNext() ) {
-            int v = iter.next();
+            int v = iter.next().intValue();
 
             if( firstValue == null ) {
                 firstValue = v;
